+++ CHANGELOG.md
@@ -10,6 +10,18 @@
 
 * Run bundle install from the Gemfile directory and support context-aware execution
 
+## [0.13.0](https://github.com/contriboss/gemfile-go/compare/v0.12.0...v0.13.0) (2026-02-15)
+
+
+### Features
+
+* Fallback to bundle install when lockfile is missing ([0232609](https://github.com/contriboss/gemfile-go/commit/023260960b8958ec083b505ae0d4b5b99671f348))
+
+
+### Miscellaneous Chores
+
+* release 0.13.0 ([0232609](https://github.com/contriboss/gemfile-go/commit/023260960b8958ec083b505ae0d4b5b99671f348))
+
 ## [0.12.0](https://github.com/contriboss/gemfile-go/compare/v0.11.0...v0.12.0) (2026-02-14)